trying to fix delete route, apparently its the wrong route

+++ controllers/userController.js
@@ -1,3 +1,5 @@
+const req = require('express/lib/request');
+const res = require('express/lib/response');
 const {
     User
 } = require('../models');
@@ -51,36 +53,16 @@ module.exports = {
             .catch((err) => {
                 console.log(err);
                 res.status(500).json(err);
-            })
+            });
     },
-    deleteUser(req, res) {
-        User.findOneAndRemove({
-                _id: req.params.userId
-            })
-            .then((user) =>
-                !user ?
-                res.status(404).json({
-                    message: 'No User with this id!'
-                }) :
-                User.findOneAndUpdate({
-                    users: req.params.userId
-                }, {
-                    $pull: {
-                        users: req.params.userId
-                    }
-                }, {
-                    new: true
-                })
-            )
-            .then((user) =>
-                !user ?
-                res.status(404).json({
-                    message: ' no user with this id!',
-                }) :
-                res.json({
-                    message: 'user successfully deleted!'
-                })
-            )
-            .catch((err) => res.status(500).json(err));
-    },
+    deleteUser(req, res){
+        User.findOneAndRemove({ _id: req.params.userId})
+        .then((user)=>
+        !user
+        ? res.status(404).json({message: "No user with this id!"})
+        : res.json(user)
+        )
+        .catch ((err)=> res.status(500).json(err));
+      },
+  
 };
